* Fixed type for time values in the kernel
* Use gets to get a line from stdin for now. FIXME: UNSAFE!

+++ DayOSKernel/hal.x86/include/idt.h
@@ -50,6 +50,6 @@ int registerHandlerProcess(int irqno, uint32_t pid);
 int resetHandlerProcess(int irqno, uint32_t pid);
 void removeHandlerProcess(uint32_t pid);
 
-long unsigned int getTickCount();
+unsigned long long getTickCount();
 
 #endif
